requirements.txt
exceptiongroup==1.2.1
iniconfig==2.0.0
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
pytest==8.2.0
scipy==1.13.0
svgpathtools==1.6.1
svgwrite==1.4.3
tomli==2.0.1
